Rename new channel to Cuda Xdtt.
[ghstack-poisoned]
